Implemented a dedicated section in the example network callback file `mqttnet.c` for Microchip Harmony. Updated the MPLABX project and Harmony configuration to use PIC32MZ2048EFM144 and PIC32MZ Ethernet Starter Kit. Tested on MPLABX 4.2 and Harmony v2.06.
