--- IDE/Microchip-Harmony/mqttnet_harmony.c
@@ -1,361 +1,0 @@
-/* mqttnet_harmony.c
- *
- * Copyright (C) 2006-2018 wolfSSL Inc.
- *
- * This file is part of wolfMQTT.
- *
- * wolfMQTT is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * wolfMQTT is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
- */
-
-/* Include the autoconf generated config.h */
-#ifdef HAVE_CONFIG_H
-    #include <config.h>
-#endif
-
-#include "wolfmqtt/mqtt_client.h"
-#include "examples/mqttnet.h"
-#include "examples/mqttexample.h"
-
-/* Simplified Microchip Harmony wolfMQTT Network Callback Example */
-
-/* Harmony requires non-blocking for all operations because its design is a
-   single thread application in big loop */
-
-#ifdef MICROCHIP_MPLAB_HARMONY
-
-#ifndef WOLFMQTT_NONBLOCK
-    #error wolfMQTT must be built with WOLFMQTT_NONBLOCK defined for Harmony
-#endif
-
-#include "app.h"
-#include "system_config.h"
-#include "tcpip/tcpip.h"
-#include <sys/errno.h>
-#include <errno.h>
-
-struct timeval {
-    int tv_sec;
-    int tv_usec;
-};
-
-#define SOCKET_INVALID (-1)
-#define SO_ERROR 0
-#define SOERROR_T uint8_t
-#undef  FD_ISSET
-#define FD_ISSET(f1, f2) (1==1)
-#define SOCK_CLOSE      closesocket
-
-/* Setup defaults */
-#ifndef SOCKET_T
-    #define SOCKET_T        int
-#endif
-#ifndef SOERROR_T
-    #define SOERROR_T       int
-#endif
-#ifndef SELECT_FD
-    #define SELECT_FD(fd)   ((fd) + 1)
-#endif
-#ifndef SOCKET_INVALID
-    #define SOCKET_INVALID  ((SOCKET_T)0)
-#endif
-#ifndef SOCK_CONNECT
-    #define SOCK_CONNECT    connect
-#endif
-#ifndef SOCK_SEND
-    #define SOCK_SEND(s,b,l,f) send((s), (b), (size_t)(l), (f))
-#endif
-#ifndef SOCK_RECV
-    #define SOCK_RECV(s,b,l,f) recv((s), (b), (size_t)(l), (f))
-#endif
-#ifndef SOCK_CLOSE
-    #define SOCK_CLOSE      close
-#endif
-#ifndef SOCK_ADDR_IN
-    #define SOCK_ADDR_IN    struct sockaddr_in
-#endif
-#ifdef SOCK_ADDRINFO
-    #define SOCK_ADDRINFO   struct addrinfo
-#endif
-
-
-/* Local context for Net callbacks */
-typedef enum {
-    SOCK_BEGIN = 0,
-    SOCK_CONN,
-} NB_Stat;
-
-typedef struct _SocketContext {
-    SOCKET_T fd;
-    NB_Stat stat;
-    SOCK_ADDR_IN addr;
-} SocketContext;
-
-
-/* Private functions */
-static int NetConnect(void *context, const char* host, word16 port,
-    int timeout_ms)
-{
-    SocketContext *sock = (SocketContext*)context;
-    int type = SOCK_STREAM;
-    int rc = MQTT_CODE_ERROR_NETWORK;
-    SOERROR_T so_error = 0;
-    struct addrinfo *result = NULL;
-    struct addrinfo hints;
-    struct hostent *hostInfo;
-
-    /* Get address information for host and locate IPv4 */
-    switch(sock->stat) {
-        case SOCK_BEGIN:
-        {
-            XMEMSET(&hints, 0, sizeof(hints));
-            hints.ai_family = AF_INET;
-            hints.ai_socktype = SOCK_STREAM;
-            hints.ai_protocol = IPPROTO_TCP;
-
-            XMEMSET(&sock->addr, 0, sizeof(sock->addr));
-            sock->addr.sin_family = AF_INET;
-
-            hostInfo = gethostbyname((char *)host);
-            if (hostInfo != NULL) {
-                sock->addr.sin_port = port; /* htons(port); */
-                sock->addr.sin_family = AF_INET;
-                XMEMCPY(&sock->addr.sin_addr.S_un,
-                            *(hostInfo->h_addr_list), sizeof(IPV4_ADDR));
-            }
-            else {
-                return MQTT_CODE_CONTINUE;
-            }
-
-            /* Default to error */
-            rc = MQTT_CODE_ERROR_NETWORK;
-
-            /* Create socket */
-            sock->fd = socket(sock->addr.sin_family, type, 0);
-            if (sock->fd == SOCKET_INVALID)
-                goto exit;
-
-            sock->stat = SOCK_CONN;
-
-            FALL_THROUGH;
-        }
-
-        case SOCK_CONN:
-        {
-            /* Start connect */
-            rc = connect(sock->fd, (struct sockaddr*)&sock->addr, sizeof(sock->addr));
-            if (rc) {
-                /* Check for error */
-                if (errno == EINPROGRESS) {
-                    rc = MQTT_CODE_CONTINUE;
-                }
-            }
-            break;
-        }
-
-        default:
-            rc = MQTT_CODE_ERROR_BAD_ARG;
-            break;
-    } /* switch */
-
-    (void)timeout_ms;
-
-exit:
-    /* Show error */
-    if (rc != 0) {
-        PRINTF("NetConnect: Rc=%d, SoErr=%d", rc, so_error);
-    }
-
-    return rc;
-}
-
-static int NetWrite(void *context, const byte* buf, int buf_len,
-    int timeout_ms)
-{
-    SocketContext *sock = (SocketContext*)context;
-    int rc;
-    SOERROR_T so_error = 0;
-
-    if (context == NULL || buf == NULL || buf_len <= 0) {
-        return MQTT_CODE_ERROR_BAD_ARG;
-    }
-
-    rc = (int)SOCK_SEND(sock->fd, buf, buf_len, 0);
-    if (rc == -1) {
-        /* Get error */
-        socklen_t len = sizeof(so_error);
-        getsockopt(sock->fd, SOL_SOCKET, SO_ERROR, &so_error, &len);
-        if (so_error == 0) {
-            rc = 0; /* Handle signal */
-        }
-        else {
-            if (so_error == EWOULDBLOCK || so_error == EAGAIN) {
-                return MQTT_CODE_CONTINUE;
-            }
-
-            rc = MQTT_CODE_ERROR_NETWORK;
-            PRINTF("NetWrite: Error %d", so_error);
-        }
-    }
-
-    (void)timeout_ms;
-
-    return rc;
-}
-
-static int NetRead(void *context, byte* buf, int buf_len,
-    int timeout_ms)
-{
-    SocketContext *sock = (SocketContext*)context;
-    int rc = MQTT_CODE_ERROR_NETWORK, timeout = 0;
-    SOERROR_T so_error = 0;
-    int bytes = 0;
-
-    if (context == NULL || buf == NULL || buf_len <= 0) {
-        return MQTT_CODE_ERROR_BAD_ARG;
-    }
-
-    /* Loop until buf_len has been read, error or timeout */
-    while (bytes < buf_len) {
-
-        /* Try and read number of buf_len provided,
-            minus what's already been read */
-        rc = (int)SOCK_RECV(sock->fd,
-                       &buf[bytes],
-                       buf_len - bytes,
-                       0);
-        if (rc <= 0) {
-            rc = MQTT_CODE_ERROR_NETWORK;
-            goto exit; /* Error */
-        }
-        else {
-            bytes += rc; /* Data */
-        }
-
-        /* non-blocking should always exit loop */
-        break;
-    } /* while */
-
-exit:
-
-    if (rc == 0 && timeout) {
-        rc = MQTT_CODE_ERROR_TIMEOUT;
-    }
-    else if (rc < 0) {
-        /* Get error */
-        socklen_t len = sizeof(so_error);
-        getsockopt(sock->fd, SOL_SOCKET, SO_ERROR, &so_error, &len);
-
-        if (so_error == 0) {
-            rc = 0; /* Handle signal */
-        }
-        else {
-            if (so_error == EWOULDBLOCK || so_error == EAGAIN) {
-                return MQTT_CODE_CONTINUE;
-            }
-
-            rc = MQTT_CODE_ERROR_NETWORK;
-            PRINTF("NetRead: Error %d", so_error);
-        }
-    }
-    else {
-        rc = bytes;
-    }
-
-    (void)timeout_ms;
-
-    return rc;
-}
-
-static int NetDisconnect(void *context)
-{
-    SocketContext *sock = (SocketContext*)context;
-    if (sock) {
-        if (sock->fd != SOCKET_INVALID) {
-            SOCK_CLOSE(sock->fd);
-            sock->fd = -1;
-        }
-
-        sock->stat = SOCK_BEGIN;
-    }
-    return 0;
-}
-
-
-/* Public Functions */
-int MqttClientNet_Init(MqttNet* net)
-{
-    static IPV4_ADDR    dwLastIP[2] = { {-1}, {-1} };
-    IPV4_ADDR           ipAddr;
-    int Dummy;
-    int nNets;
-    int i;
-    SYS_STATUS          stat;
-    TCPIP_NET_HANDLE    netH;
-
-    stat = TCPIP_STACK_Status(sysObj.tcpip);
-    if (stat < 0) {
-        return MQTT_CODE_CONTINUE;
-    }
-
-    nNets = TCPIP_STACK_NumberOfNetworksGet();
-    for (i = 0; i < nNets; i++) {
-        netH = TCPIP_STACK_IndexToNet(i);
-        ipAddr.Val = TCPIP_STACK_NetAddress(netH);
-        if (ipAddr.v[0] == 0) {
-            return MQTT_CODE_CONTINUE;
-        }
-        if (dwLastIP[i].Val != ipAddr.Val) {
-            dwLastIP[i].Val = ipAddr.Val;
-            PRINTF("%s", TCPIP_STACK_NetNameGet(netH));
-            PRINTF(" IP Address: ");
-            PRINTF("%d.%d.%d.%d\n", ipAddr.v[0], ipAddr.v[1], ipAddr.v[2], ipAddr.v[3]);
-        }
-    }
-
-    if (net) {
-        XMEMSET(net, 0, sizeof(MqttNet));
-        net->connect = NetConnect;
-        net->read = NetRead;
-        net->write = NetWrite;
-        net->disconnect = NetDisconnect;
-        net->context = (SocketContext *)WOLFMQTT_MALLOC(sizeof(SocketContext));
-        if (net->context == NULL) {
-            return MQTT_CODE_ERROR_MEMORY;
-        }
-        XMEMSET(net->context, 0, sizeof(SocketContext));
-
-        ((SocketContext*)(net->context))->stat = SOCK_BEGIN;
-    }
-
-    return MQTT_CODE_SUCCESS;
-}
-
-int MqttClientNet_DeInit(MqttNet* net)
-{
-    if (net) {
-        if (net->context) {
-            WOLFMQTT_FREE(net->context);
-        }
-        XMEMSET(net, 0, sizeof(MqttNet));
-    }
-    return 0;
-}
-
-#else
-
-    #error This examplei s designed to work with Microchip Harmony only
-    /* see the complete mqttnet.c examples in ./examples/mqttnet.c */
-
-#endif /* MICROCHIP_MPLAB_HARMONY */
